feat: allow setting twingate-enabled
When twingate-enabled is set to true, a twingate client is started using
a pre-defined service key for a "github_actions" service account.

No proxies are required with Twingate (unlike Tailscale) because DNS
records get overridden for select endpoints. The service account is
currently setup to route traffic through Twingate for cap1 DNS names
{155210249240,168414722295}.cap1.observeinc.com.