feat: synchronize github secrets to secretsmanager (#124)
It is more convenient to manage secrets in github than to interact with
AWS secretsmanager. Unfortunately, doing so naively would make it
impossible for an operator to run terraform manually if a complex state
operation were to be necessary. As a workaround, we can ensure that
secrets with prefix `X_` are synchronized to secretsmanager on every
workflow run.